Allow overriding field serializer via include.
Previous behavior was non-determinstic: sticking include members which
were tuples in a set with default_fields did not replace the default
field name. For example if the model has a field "myfield" and include
has ('myfield', MyFieldSerializer), the resulting set would be
('myfield', ('myfield', MyFieldSerializer)). The order of that is not
guaranteed.